Add rough click functionality
